pyproject.toml
[project]
name = "anki-skill"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "genanki>=0.13.1",
    "openai>=2.33.0",
    "python-dotenv>=1.2.2",
]
